% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantha.R
\name{bantha}
\alias{bantha}
\title{Bayesian Network Thresholding Heuristic Algorithm (BANTHA)}
\usage{
bantha(samples, a = 1, n_candidates = 0, n_cores = 0)
}
\arguments{
\item{samples}{An object of class \sQuote{array} containing posterior samples
from a Bayesian network distribution. Each array element encodes one
network as a binary adjacency matrix, with nodes in the rows and columns.
A value of 1 in cell (i, j) indicates a directed edge from node i to node j.}

\item{a}{A numeric scalar for the cost parameter of generalized Hamming
distance used in GSH loss.  The other cost parameter, \eqn{b}, is equal to
\eqn{2 - a}.}

\item{n_candidates}{The number of starting states to use in calculating the best
matrix estimate. Using all samples as starting states is recommended is
indicated by a value of zero.}

\item{n_cores}{The number of CPU cores to use, i.e., the number of
simultaneous calculations at any given time. A value of zero indicates to
use all cores on the system.}
}
\value{
A Bayes estimate in the form of an adjacency matrix found by minimizing
the Monte Carlo approximation of the expected GSH loss using the available
samples.
}
\description{
This function provides a Bayesian network to summarize a network distribution
using the BANTHA method (Bailey, Dahl, Andros 2024).
}
\examples{
data(choi25)
bantha(choi25, a = 1)

}
\references{
E. Bailey, D. B. Dahl, J. Andros (2025).
Structure Learning of Bayesian Networks from Posterior Sample Inference.
Unpublished manuscript. Available upon request from the authors.
}
