% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_expected_gsh_loss.R
\name{compute_expected_gsh_loss}
\alias{compute_expected_gsh_loss}
\title{Compute Expected Generalized Structural Hamming (GSH) Loss}
\usage{
compute_expected_gsh_loss(network, gsh)
}
\arguments{
\item{network}{A Bayesian network represented by an adjacency matrix.}

\item{gsh}{A pointer to an initialized GSH structure containing the posterior
probabilities of edges between each pair of nodes, which can be found using
the \code{\link{initialized_expected_gsh_loss}} function.}
}
\value{
The expected GSH loss for the given Bayesian network based on the posterior
posterior probabilities of edges.
}
\description{
This function computes the expected GSH loss (Bailey, Dahl, Andros, 2024)
for a given adjacency matrix.
}
\examples{
data(choi25)
gsh <- initialized_expected_gsh_loss(choi25, a = 1)
compute_expected_gsh_loss(choi25[,,1], gsh)
}
\references{
Bailey, E., Dahl, D. B., Andros, J. (2024).
Structure Learning of Bayesian Networks from Posterior Sample Inference.
Unpublished manuscript. Available upon request from the authors.
}
