% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsh_loss_builder_process.R
\name{gsh_loss_builder_process}
\alias{gsh_loss_builder_process}
\title{Add Bayesian networks to GSH builder structure}
\usage{
gsh_loss_builder_process(x, gsh_loss_builder)
}
\arguments{
\item{x}{A network as an adjacency matrix.}

\item{gsh_loss_builder}{A pointer to a GSH builder structure.}
}
\description{
Add a Bayesian network represented by an adjacency matrix to an existing GSH
builder structure
}
\examples{
data(choi25)
gsh_builder <- gsh_loss_builder_new(ncol(choi25), nrow(choi25), 4, a = 1)
gsh_loss_builder_process(choi25[,,1], gsh_builder)

}
\references{
E. Bailey, D. B. Dahl, J. Andros (2025).
Structure Learning of Bayesian Networks from Posterior Sample Inference.
Unpublished manuscript. Available upon request from the authors.
}
